requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
pillow==10.0.1
torch>=2.2.0
torchvision>=0.17.0
